requirements.txt
# Core API
fastapi==0.115.12
uvicorn==0.27.1

# Database
SQLAlchemy==2.0.41
alembic==1.13.1

# Validation & settings
pydantic==2.10.6
pydantic-settings==2.8.1

# Templates & forms (HTML captive portal)
jinja2==3.1.6
python-multipart==0.0.19

# MikroTik integration
routeros-api==0.21.0

# Payments (future-ready)
paynow==1.0.8

# Background jobs (future)
celery==5.5.3

# Utilities
qrcode[pil]==7.4.2
Pillow==10.2.0

# Dev / tooling
setuptools==75.3.3
wheel==0.45.1

reportlab